+++ src/DataManager.ts
@@ -1,5 +1,4 @@
 import * as validator from 'json-schema-remote';
-import * as EventSource from 'eventsource/lib/eventsource-polyfill';
 
 import Core, { environment, options } from './Core';
 import DataManagerResource from './resources/datamanager/DataManagerResource';
@@ -9,15 +8,15 @@ import DMStatsResource from './resources/datamanager/DMStatsResource';
 import TemplateList from './resources/datamanager/TemplateList';
 import TemplateResource from './resources/datamanager/TemplateResource';
 import { filterOptions } from './resources/ListResource';
-import { get, getUrl, optionsToQuery, post, superagentGet } from './helper';
-import TokenStoreFactory, { TokenStore } from './TokenStore';
+import { get, getHistory, optionsToQuery, post, superagentGet } from './helper';
+
+declare const EventSource: any;
 
 validator.setLoggingFunction(() => {
 });
 
 const environmentSymbol = Symbol.for('environment');
 const relationsSymbol = Symbol.for('relations');
-const tokenStoreSymbol = Symbol.for('tokenStore');
 
 const urls = {
   live: 'https://datamanager.entrecode.de/',
@@ -187,7 +186,7 @@ export default class DataManager extends Core {
    * @param {filterOptions | any} options The filter options
    * @return {Promise<EventSource>} The created EventSource.
    */
-  newHistory(options?: filterOptions | any) {
+  newHistory(options?: filterOptions): Promise<any> {
     return Promise.resolve()
     .then(() => this.follow('ec:history'))
     .then(request => {
@@ -197,31 +196,7 @@ export default class DataManager extends Core {
         request.withTemplateParameters(optionsToQuery(options));
       }
 
-      return getUrl(this[environmentSymbol], request)
-    })
-    .then((url) => {
-      const eventSourceInitDict = {
-        headers: {},
-      };
-
-      const store = this[tokenStoreSymbol];
-      let secondStore: TokenStore;
-      if (!store.hasToken()) {
-        // when no token is present see if we have a public environment (with shortID)
-        // if so look in second store
-        const result = /^(live|stage|nightly|develop|test)[A-Fa-f0-9]{8}$/.exec(this[environmentSymbol]);
-        if (result) {
-          secondStore = TokenStoreFactory(<environment>result[1]);
-        }
-      }
-
-      if (store.hasToken()) {
-        eventSourceInitDict.headers['Authorization'] = `Bearer ${store.getToken()}`;
-      } else if (secondStore && secondStore.hasToken()) {
-        eventSourceInitDict.headers['Authorization'] = `Bearer ${secondStore.getToken()}`;
-      }
-
-      return new EventSource(url, eventSourceInitDict);
+      return getHistory(this[environmentSymbol], request)
     });
   }